Code cleaning.
