+++ app/GoodReads/credits.txt
