+++ app/error_messages.py
@@ -4,7 +4,7 @@ class ValidationMessages (object):
     SUCCESS = 100
     NO_ACTIVE_OR_PAUSED_READING_WAS_FOUND = 101
     NO_BOOK_WAS_FOUND_WITH_THE_GIVEN_ISBN_CODE = 102
-    BOOK_HAS_ALREADY_BEEN_ADDED_TO_THE_BOOK_SHELF = 103
+    BOOK_HAS_ALREADY_BEEN_ADDED_TO_BOOK_SHELF = 103
